hw-mgmt: patches: v4.19: Support for multi PWM
Add additional PWM attributes in order to support the systems, which
can be equipped with up-to four PWM controllers. System capability of
additional PWM support is validated through the reading of relevant
registers.

Add support for additional cooling devices in order to support the
systems, which can be equipped with up-to four PWM controllers.

Add new registers to support systems with multiply cooling devices.
Some new systems can support up-to four cooling devices. This
capability is detected according to the registers initial setting.

Signed-off-by: Vadim Pasternak <vadimp@nvidia.com>